refactor(sequencer): extract l1 origin module
Amp-Thread-ID: https://ampcode.com/threads/T-019fec00-7145-73ea-8665-a00aadbb36dd
