+++ crates/consensus/service/src/actors/sequencer/l1_origin/mod.rs
@@ -1,0 +1,9 @@
+//! L1 origin selection and provider implementations for the sequencer.
+
+mod provider;
+pub use provider::{DelayedL1OriginSelectorProvider, L1OriginSelectorProvider};
+
+mod selector;
+#[cfg(test)]
+pub use selector::MockOriginSelector;
+pub use selector::{L1OriginSelector, L1OriginSelectorError, OriginSelector};
